refactor: constants 파일 분리
