+++ src/app/(route)/service-introduction/_constants/SERVICE_BENEFITS.ts
@@ -1,0 +1,20 @@
+export const serviceBenefits = [
+  {
+    content: "건강한 스포츠 토론을 즐기고 싶은 분!",
+    imgLink: "Service_benefit_1",
+  },
+  {
+    content: "경기 중 실시간으로 의견을 나누고 싶은 분!",
+    reverse: true,
+    imgLink: "Service_benefit_2",
+  },
+  {
+    content: "관심 있는 팀과 선수 소식을 빠르게 공유!",
+    imgLink: "Service_benefit_3",
+  },
+  {
+    content: "팬들과 함께 승리의 순간을 공유하는 즐거움!",
+    reverse: true,
+    imgLink: "Service_benefit_4",
+  },
+];
