+++ src/app/(route)/service-introduction/_constants/SERVICE_ITEMS.ts
@@ -1,0 +1,35 @@
+export const serviceItems = [
+  {
+    imgUrl: "Service_feature_1",
+    title: "경기중계와 실시간 채팅",
+    content: "팬들과 함께 경기 라인업, 기록 등을 보며\n 즐겁게 소통하는 채팅",
+  },
+  {
+    imgUrl: "Service_feature_2",
+    title: "안전한 커뮤니티 운영",
+    content:
+      "신고 시스템, 비속어 필터링으로 지속관리\n 관리자의 모니터링까지",
+  },
+  {
+    imgUrl: "Service_feature_3",
+    title: "건강한 응원 문화",
+    content:
+      "비방,악성 댓글 없는 클린지향주의\n 서로를 존중할 자세가 있다면 컴온!",
+  },
+  {
+    imgUrl: "Service_feature_4",
+    title: "다양한 카테고리 게시판",
+    content:
+      "E스포츠, 축구, 야구 등의 다양한 허브\n 니즈에 따라 추가제공 가능",
+  },
+  {
+    imgUrl: "Service_feature_5",
+    title: "승부예측",
+    content: "과연 누가 이길까?\n 나의 팀을 응원하고 예측하는 재미!",
+  },
+  {
+    imgUrl: "Service_feature_6",
+    title: "실시간 뉴스기사",
+    content: "최신 뉴스 기사와 함께 토론해볼수 있는\n 이야기 소통 공간 제공",
+  },
+];
